+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+- allow for adaptors closing the file handle themselves
+- fixed broken unit-tests changed MimeType detection of [Flysystem 3.0.16](https://github.com/thephpleague/flysystem/compare/3.0.15...3.0.16)
+
 ## [1.0.3] - 2022-01-21
 ### Added
 - allow flysystem v3
@@ -22,10 +26,10 @@
 - ...
 --->
 
+[Unreleased]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.3...HEAD
 [1.0.3]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/m2mtech/flysystem-stream-wrapper/releases/tag/v1.0.0
 <!---
-[Unreleased]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.1...HEAD
 --->