+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## [Unreleased]
+## [1.1.0] - 2022-05-05
 - allow for adaptors closing the file handle themselves
-- fixed broken unit-tests changed MimeType detection of [Flysystem 3.0.16](https://github.com/thephpleague/flysystem/compare/3.0.15...3.0.16)
+- fixed broken directory detection caused by changed MimeType detection of [Flysystem 3.0.16](https://github.com/thephpleague/flysystem/compare/3.0.15...3.0.16)
 - allow to ignore visibility errors
 
 ## [1.0.3] - 2022-01-21
@@ -27,10 +27,11 @@
 - ...
 --->
 
-[Unreleased]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.3...HEAD
+[1.1.0]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.3...v1.1.0
 [1.0.3]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/m2mtech/flysystem-stream-wrapper/releases/tag/v1.0.0
 <!---
+[Unreleased]: https://github.com/m2mtech/flysystem-stream-wrapper/compare/v1.0.3...HEAD
 --->